Refactoring of integration tests
