+++ integration_tests/cairo1_test.go
@@ -1,0 +1,391 @@
+package integrationtests
+
+import (
+	"flag"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+	"sync"
+	"testing"
+	"time"
+
+	"github.com/consensys/gnark-crypto/ecc/stark-curve/fp"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func runAndTestFile(t *testing.T, path string, name string, benchmarkMap map[string][]int, benchmark bool, errorExpected bool, inputArgs string, proofmode bool) {
+	t.Logf("testing: %s\n", path)
+	compiledOutput, err := compileCairoCode(path)
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	layout := getLayoutFromFileName(path)
+
+	elapsedGo, traceFile, memoryFile, _, err := runVmCairo1(compiledOutput, layout, inputArgs, proofmode)
+	if errorExpected {
+		assert.Error(t, err, path)
+		writeToFile(path)
+		return
+	} else {
+		if err != nil {
+			t.Error(err)
+			writeToFile(path)
+			return
+		}
+	}
+
+	elapsedRs, rsTraceFile, rsMemoryFile, err := runRustVmCairo1(path, layout, inputArgs, proofmode)
+	if errorExpected {
+		// we let the code go on so that we can check if the go vm also raises an error
+		assert.Error(t, err, path)
+		return
+	} else {
+		if err != nil {
+			t.Error(err)
+			writeToFile(path)
+			return
+		}
+	}
+
+	trace, memory, err := decodeProof(traceFile, memoryFile)
+	if err != nil {
+		t.Error(err)
+		writeToFile(path)
+		return
+	}
+	rsTrace, rsMemory, err := decodeProof(rsTraceFile, rsMemoryFile)
+	if err != nil {
+		t.Error(err)
+		writeToFile(path)
+		return
+	}
+
+	if !assert.Equal(t, rsTrace, trace) {
+		t.Logf("rstrace:\n%s\n", traceRepr(rsTrace))
+		t.Logf("trace:\n%s\n", traceRepr(trace))
+		writeToFile(path)
+	}
+	if !assert.Equal(t, rsMemory, memory) {
+		t.Logf("rsmemory;\n%s\n", memoryRepr(rsMemory))
+		t.Logf("memory;\n%s\n", memoryRepr(memory))
+		writeToFile(path)
+	}
+
+	if benchmark {
+		benchmarkMap[name] = []int{int(elapsedGo.Milliseconds()), int(elapsedRs.Milliseconds())}
+	}
+}
+
+func compareWithStarkwareRunner(t *testing.T, path string, errorExpected bool, inputArgs string) {
+	t.Logf("testing (comapre with cairo runner): %s\n", path)
+
+	runnerMemory, err := runCairoRunner(path)
+	if err != nil {
+		t.Error(err)
+		writeToFile(path)
+		return
+	}
+
+	compiledOutput, err := compileCairoCode(path)
+	if err != nil {
+		t.Error(err)
+		return
+	}
+
+	_, traceFile, memoryFile, _, err := runVmCairo1(compiledOutput, "all_cairo", inputArgs, false)
+	if errorExpected {
+		assert.Error(t, err, path)
+		writeToFile(path)
+		return
+	} else {
+		if err != nil {
+			t.Error(err)
+			writeToFile(path)
+			return
+		}
+	}
+	_, memory, err := decodeProof(traceFile, memoryFile)
+	if err != nil {
+		t.Error(err)
+		writeToFile(path)
+		return
+	}
+
+	assert.Equal(t, runnerMemory, memory)
+}
+
+var cairobench = flag.Bool("cairobench", false, "run integration tests and generate benchmarks file")
+
+func TestCairoFiles(t *testing.T) {
+	file, err := os.OpenFile(whitelistFile, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0644)
+	if err != nil {
+		panic(fmt.Errorf("failed to open file: %w", err))
+	}
+
+	file.Close()
+
+	roots := []string{
+		// "./cairo_1_programs/",
+		// "./cairo_1_programs/dict_non_squashed",
+		// "./cairo_1_programs/with_input",
+		// "./cairo_1_programs/serialized_output",
+	}
+
+	inputArgsMap := map[string]string{
+		"cairo_1_programs/with_input/array_input_sum__small.cairo":         "2 [111 222 333] 1 [444 555 666 777]",
+		"cairo_1_programs/with_input/array_length__small.cairo":            "[1 2 3 4 5 6] [7 8 9 10]",
+		"cairo_1_programs/with_input/branching.cairo":                      "123",
+		"cairo_1_programs/with_input/dict_with_input__small.cairo":         "[1 2 3 4]",
+		"cairo_1_programs/with_input/tensor__small.cairo":                  "[1 4] [1 5]",
+		"cairo_1_programs/with_input/proofmode__small.cairo":               "[1 2 3 4 5]",
+		"cairo_1_programs/with_input/proofmode_with_builtins__small.cairo": "[1 2 3 4 5]",
+		"cairo_1_programs/with_input/proofmode_segment_arena__small.cairo": "[1 2 3 4 5]",
+	}
+
+	// filter is for debugging purposes
+	filter := Filter{}
+	filter.init()
+
+	benchmarkMap := make(map[string][]int)
+
+	sem := make(chan struct{}, 5) // semaphore to limit concurrency
+	var wg sync.WaitGroup         // WaitGroup to wait for all goroutines to finish
+
+	for _, root := range roots {
+		testFiles, err := os.ReadDir(root)
+		require.NoError(t, err)
+
+		for _, dirEntry := range testFiles {
+			if dirEntry.IsDir() || isGeneratedFile(dirEntry.Name()) {
+				continue
+			}
+
+			name := dirEntry.Name()
+			path := filepath.Join(root, name)
+
+			errorExpected := false
+			if name == "range_check__small.cairo" {
+				errorExpected = true
+			}
+			if !filter.filtered(name) {
+				continue
+			}
+			inputArgs := inputArgsMap[path]
+			// we run tests concurrently if we don't need benchmarks
+			if !*cairobench {
+				sem <- struct{}{} // acquire a semaphore slot
+				wg.Add(1)
+
+				go func(path, name, inputArgs string) {
+					defer wg.Done()
+					defer func() { <-sem }() // release the semaphore slot when done
+					runAndTestFile(t, path, name, benchmarkMap, *cairobench, errorExpected, inputArgs, false)
+					if strings.Contains(path, "proofmode") || strings.Contains(path, "serialized_output/with_input") {
+						runAndTestFile(t, path, name, benchmarkMap, *cairobench, errorExpected, inputArgs, true)
+					}
+					if !strings.Contains(path, "with_input") {
+						compareWithStarkwareRunner(t, path, errorExpected, inputArgs)
+					}
+				}(path, name, inputArgs)
+			} else {
+				runAndTestFile(t, path, name, benchmarkMap, *cairobench, errorExpected, inputArgs, true)
+				if strings.Contains(path, "proofmode") || strings.Contains(path, "serialized_output/with_input") {
+					runAndTestFile(t, path, name, benchmarkMap, *cairobench, errorExpected, inputArgs, true)
+				}
+				if !strings.Contains(path, "with_input") {
+					compareWithStarkwareRunner(t, path, errorExpected, inputArgs)
+				}
+			}
+		}
+	}
+
+	wg.Wait() // wait for all goroutines to finish
+
+	for _, root := range roots {
+		clean(root)
+	}
+
+	if *cairobench {
+		WriteBenchMarksToFile(benchmarkMap)
+	}
+}
+
+// given a path to a cairo zero file, it compiles it
+// and return the compilation path
+func compileCairoCode(path string) (string, error) {
+	if filepath.Ext(path) != ".cairo" {
+		return "", fmt.Errorf("compiling non cairo file: %s", path)
+	}
+	compiledOutput := swapExtenstion(path, compiledSuffix)
+
+	var cliCommand string
+	var args []string
+
+	sierraOutput := swapExtenstion(path, sierraSuffix)
+	cliCommand = "../rust_vm_bin/starkware/starkware/cairo-compile"
+	args = []string{
+		"--single-file",
+		path,
+		sierraOutput,
+		"--replace-ids",
+	}
+
+	cmd := exec.Command(cliCommand, args...)
+
+	res, err := cmd.CombinedOutput()
+	if err != nil {
+		return "", fmt.Errorf(
+			"%s %s: %w\n%s", cliCommand, path, err, string(res),
+		)
+	}
+
+	cliCommand = "../rust_vm_bin/scj/scj/sierra-compile-json"
+	args = []string{
+		sierraOutput,
+		compiledOutput,
+	}
+
+	cmd = exec.Command(cliCommand, args...)
+
+	res, err = cmd.CombinedOutput()
+	if err != nil {
+		return "", fmt.Errorf(
+			"%s %s: %w\n%s", cliCommand, path, err, string(res),
+		)
+	}
+
+	return compiledOutput, nil
+}
+
+// given a path to a compiled cairo1 file, execute it using the
+// rust vm and return the trace and memory files location
+func runRustVmCairo1(path, layout string, inputArgs string, proofmode bool) (time.Duration, string, string, error) {
+	traceOutput := swapExtenstion(path, rsTraceSuffix)
+	memoryOutput := swapExtenstion(path, rsMemorySuffix)
+
+	args := []string{
+		path,
+		"--trace_file",
+		traceOutput,
+		"--memory_file",
+		memoryOutput,
+		"--layout",
+		layout,
+		"--args",
+		inputArgs,
+	}
+
+	if proofmode {
+		args = append(args, "--proof_mode")
+	}
+
+	cmd := exec.Command("./../rust_vm_bin/lambdaclass/lambdaclass/cairo1-run", args...)
+
+	start := time.Now()
+
+	res, err := cmd.CombinedOutput()
+
+	elapsed := time.Since(start)
+
+	if err != nil {
+		return 0, "", "", fmt.Errorf(
+			"%s %s: %w\n%s", "./../rust_vm_bin/lambdaclass/lambdaclass/cairo1-run", path, err, string(res),
+		)
+	}
+
+	return elapsed, traceOutput, memoryOutput, nil
+}
+
+// given a path to a compiled cairo1 file, execute
+// it using our vm
+func runVmCairo1(path, layout string, inputArgs string, proofmode bool) (time.Duration, string, string, string, error) {
+	traceOutput := swapExtenstion(path, traceSuffix)
+	memoryOutput := swapExtenstion(path, memorySuffix)
+
+	args := []string{
+		"cairo-run",
+		"--tracefile",
+		traceOutput,
+		"--memoryfile",
+		memoryOutput,
+		"--layout",
+		layout,
+		"--available_gas",
+		"9999999",
+		"--args",
+		inputArgs,
+	}
+
+	if proofmode {
+		args = append(args, "--proofmode")
+	}
+
+	args = append(args, path)
+
+	cmd := exec.Command(
+		"../bin/cairo-vm",
+		args...,
+	)
+
+	start := time.Now()
+
+	res, err := cmd.CombinedOutput()
+
+	elapsed := time.Since(start)
+
+	if err != nil {
+		return 0, "", "", string(res), fmt.Errorf(
+			"cairo-vm run %s: %w\n%s", path, err, string(res),
+		)
+	}
+
+	return elapsed, traceOutput, memoryOutput, string(res), nil
+}
+
+func runCairoRunner(path string) ([]fp.Element, error) {
+	args := []string{
+		"--single-file",
+		path,
+		"--available-gas",
+		"9999999",
+		"--print-full-memory",
+	}
+
+	cmd := exec.Command("./../rust_vm_bin/starkware/starkware/cairo-run", args...)
+	rsOutputByte, err := cmd.CombinedOutput()
+
+	if err != nil {
+		return nil, fmt.Errorf(
+			"cairo-vm run %s: %w\n%s", path, err, string(rsOutputByte),
+		)
+	}
+	rsOutput := string(rsOutputByte)
+	// Extract memory values from output string
+	memoryStart := strings.Index(rsOutput, "Full memory: [") + 14
+	memoryEnd := strings.LastIndex(rsOutput, "]") - 2
+	if memoryStart < 14 || memoryEnd == -1 {
+		writeToFile(path)
+		return nil, fmt.Errorf("Could not find memory values in output")
+	}
+	memoryStr := rsOutput[memoryStart:memoryEnd]
+	memoryStrs := strings.Split(memoryStr, ", ")
+	// Convert strings to fp.Elements
+	runnerMemory := make([]fp.Element, 0, len(memoryStrs))
+	for _, str := range memoryStrs {
+		if str == "_" {
+			runnerMemory = append(runnerMemory, fp.Element{})
+			continue
+		}
+		elem, err := new(fp.Element).SetString(str)
+		if err != nil {
+			return nil, fmt.Errorf("Could not parse memory value: %s", str)
+		}
+		runnerMemory = append(runnerMemory, *elem)
+	}
+
+	return runnerMemory, nil
+}
